eliminate shared.go
