+++ test-integ/peering_commontopo/ac4_proxy_defaults_test.go
@@ -81,7 +81,7 @@ func (s *ac4ProxyDefaultsSuite) setup(t *testing.T, ct *commonTopo) {
 		Partition: partition,
 	}
 	client := serviceExt{
-		Service: NewFortioServiceWithDefaults(
+		Service: newFortioServiceWithDefaults(
 			clu.Datacenter,
 			clientSID,
 			func(s *topology.Service) {
@@ -107,7 +107,7 @@ func (s *ac4ProxyDefaultsSuite) setup(t *testing.T, ct *commonTopo) {
 	clientNode := ct.AddServiceNode(clu, client)
 
 	server := serviceExt{
-		Service: NewFortioServiceWithDefaults(
+		Service: newFortioServiceWithDefaults(
 			peerClu.Datacenter,
 			serverSID,
 			nil,
